Fixing wagtail migration dependencies

+++ cms/migrations/0016_change_strand_indexes_to_strandpage.py
@@ -5,6 +5,7 @@ from __future__ import unicode_literals
 from django.db import migrations, models
 from django.contrib.contenttypes.models import ContentType
 from cms.models import HomePage, IndexPage, StrandPage
+import modelcluster 
 
 def swap_types(apps, schema_editor):
     # Define our content types
@@ -13,29 +14,31 @@ def swap_types(apps, schema_editor):
 
     # Find the homepage and its indexpage children
     homepage = HomePage.objects.first()
-    index_pages = homepage.get_children().filter(content_type=index_ct).exclude(title__icontains='about')
-    for p in index_pages.all():
-        children = p.get_children()
+    if homepage:
+        index_pages = homepage.get_children().filter(content_type=index_ct).exclude(title__icontains='about')
 
-        sp = StrandPage()
-        sp.title = p.title
-        sp.specific.body = p.specific.body
-        sp.parent = HomePage
-        sp.slug = "temp"
-        sp.path = '0'
-        sp.depth = '0'
-        sp.save()
-        sp.slug = p.slug
+        for p in index_pages.all():
+            children = p.get_children()
 
-        for child in children.all():
-            c.parent = sp
-            c.save()
+            sp = StrandPage()
+            sp.title = p.title
+            sp.specific.body = p.specific.body
+            sp.parent = HomePage
+            sp.slug = "temp"
+            sp.path = '0'
+            sp.depth = '0'
+            sp.save()
+            sp.slug = p.slug
 
-        sp.path = p.path
-        sp.depth = p.depth
+            for child in children.all():
+                c.parent = sp
+                c.save()
 
-        p.delete()
-        sp.save()
+            sp.path = p.path
+            sp.depth = p.depth
+
+            p.delete()
+            sp.save()
 
 
 class Migration(migrations.Migration):
